+++ posthog/src/test/java/com/posthog/PostHogTest.kt
@@ -1045,7 +1045,7 @@ internal class PostHogTest {
     }
 
     @Test
-    fun `does not set session id when reset is called`() {
+    fun `reset session id when reset is called`() {
         val http = mockHttp()
         val url = http.url("/")
 
@@ -1069,7 +1069,8 @@ internal class PostHogTest {
         var batch = serializer.deserialize<PostHogBatchEvent>(content.reader())
 
         var theEvent = batch.batch.first()
-        assertNotNull(theEvent.properties!!["\$session_id"])
+        val currentSessionId = theEvent.properties!!["\$session_id"]
+        assertNotNull(currentSessionId)
 
         sut.reset()
 
@@ -1091,7 +1092,10 @@ internal class PostHogTest {
         batch = serializer.deserialize<PostHogBatchEvent>(content.reader())
 
         theEvent = batch.batch.first()
-        assertNull(theEvent.properties!!["\$session_id"])
+        val newSessionId = theEvent.properties!!["\$session_id"]
+        assertNotNull(newSessionId)
+
+        assertTrue(currentSessionId != newSessionId)
 
         sut.close()
     }