+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - recording: expose session id ([#166](https://github.com/PostHog/posthog-android/pull/166))
+- fix: rotate session id when reset is called ([#168](https://github.com/PostHog/posthog-android/pull/168))
 - feat: add `$is_identified` property ([#167](https://github.com/PostHog/posthog-android/pull/167))
 - fix: identify should not allow already identified users ([#167](https://github.com/PostHog/posthog-android/pull/167))